backend/requirements.txt
backcall==0.1.0
decorator==4.3.2
Django==2.1.7
django-model-utils==3.1.2
djangorestframework==3.9.1
fancycompleter==0.8
jedi==0.13.3
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.9
psycopg2==2.7.7
ptyprocess==0.6.0
Pygments==2.3.1
pytz==2018.9
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
wmctrl==0.3
gunicorn==19.9.0
whitenoise==4.1.2
django-storages==1.7.1
django-debug-toolbar==1.11
django-multiselectfield==0.1.8
boto3==1.9.122
tensorflow==1.13.2
numpy==1.16.2
django-nose==1.4.6
django-environ==0.4.5
psycopg2-binary==2.8.2
django-map-widgets==0.2.2
django-nested-admin==3.2.3
django-admin-tools==0.8.1
timeloop==1.0.2
cchardet==2.1.4
django-admin-list-filter-dropdown==1.0.3
